requirements.txt
absl-py==1.2.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
api-display-purposes==0.0.3
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.9.0
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
clarifai==2.6.2
click==8.1.7
configparser==3.8.1
cryptography==37.0.4
decorator==4.4.2
EasyProcess==1.1
emoji==1.7.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
Flask==3.1.0
frozenlist==1.4.0
future==0.18.2
googleapis-common-protos==1.56.4
grpcio==1.47.0
gunicorn==23.0.0
h11==0.13.0
huepy==1.2.1
idna==2.10
imageio==2.19.3
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instabot==0.117.0
install==1.3.5
instapy==0.6.16
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==2.6.0
lxml==4.9.3
MarkupSafe==3.0.2
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==1.0.3
multidict==6.0.4
numpy==2.2.0
openai==0.28.0
openpyxl==3.1.5
ortools==9.4.1874
outcome==1.2.0
packaging==21.3
pandas==2.1.1
pandas-stubs==2.2.3.241126
Pillow==9.1.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.0.0
plyer==2.0.0
proglog==0.1.10
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==3.0
PyYAML==6.0
regex==2022.7.9
requests==2.28.0
requests-toolbelt==0.9.1
responses==0.21.0
schedule==1.1.0
selenium==4.3.0
semantic-version==2.10.0
setuptools-rust==1.4.1
six==1.16.0
sniffio==1.2.0
somepackage==1.2.3
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
starlette==0.41.3
tomli==2.0.1
tornado==6.2
tqdm==4.67.1
trio==0.21.0
trio-websocket==0.9.2
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.34.0
webdriverdownloader==1.1.0.3
Werkzeug==3.1.3
wsproto==1.1.0
yarl==1.9.2
